Easter Egg and more precice collisions
There's an easter egg if the player decides to jump out of the window.
The hitboxes for the Enemies, Citizens, and the VIP use the ragdoll
colliders rather than the large capsule collider or sphere collider,
making them more precice and allowing for the player to shoot an enemy
behind another target, if they are a good enough shot.